Outsourcing the file logging functionality into common. Adding new common API and tests
