+++ tst/FileLoggerFunctionalityTest.cpp
@@ -38,10 +38,10 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
                                                                          &pClientCallbacks));
 
         // make sure the files dont exist
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLogIndex");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.0");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.1");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.2");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2");
 
         EXPECT_EQ(STATUS_SUCCESS, addFileLoggerPlatformCallbacksProvider(pClientCallbacks, MIN_FILE_LOGGER_STRING_BUFFER_SIZE, 5, TEST_TEMP_DIR_PATH_NO_ENDING_SEPARTOR, FALSE));
         logFunc = pClientCallbacks->logPrintFn;
@@ -49,39 +49,39 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
         logFunc(LOG_LEVEL_VERBOSE, NULL, (PCHAR) "%s", logMessage);
         logFunc(LOG_LEVEL_VERBOSE, NULL, (PCHAR) "%s", logMessage);
         // low log level logs have no effect
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(FALSE, fileFound);
 
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
         // log should not have been flushed because we havent fill out string buffer
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(FALSE, fileFound);
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
         EXPECT_EQ(FALSE, fileFound); // index file should not be there
 
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
         // log should have been flushed because the new message overflows string buffer
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
         // second log is still in buffer. Thus this log will cause the second log to be flushed.
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
         // log should have been flushed
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.1"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
         EXPECT_EQ(TRUE, fileFound); // index file should be there
 
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), TRUE, NULL, &fileIndexBufferSize));
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), TRUE, (PBYTE) fileIndexBuffer, &fileIndexBufferSize));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, NULL, &fileIndexBufferSize));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, (PBYTE) fileIndexBuffer, &fileIndexBufferSize));
         fileIndexBuffer[fileIndexBufferSize] = '\0';
         STRTOUI64(fileIndexBuffer, NULL, 10, &currentFileIndex);
         EXPECT_EQ(2, currentFileIndex); // index should be 2 since we flushed twice.
 
         EXPECT_EQ(STATUS_SUCCESS, freeCallbacksProvider(&pClientCallbacks));
 
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.2"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2"), &fileFound));
         EXPECT_EQ(TRUE, fileFound); // remaining log get flushed when callbacks are freed.
 
         MEMFREE(logMessage);
@@ -109,9 +109,9 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
                                                                          &pClientCallbacks));
 
         // make sure the files dont exist
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLogIndex");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
         for(; i < logIterationCount; ++i) {
-            SPRINTF(filePath, TEST_TEMP_DIR_PATH "kvsProducerLog.%u", i);
+            SPRINTF(filePath, TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".%u", i);
             FREMOVE(filePath);
         }
 
@@ -126,7 +126,7 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
         EXPECT_EQ(STATUS_SUCCESS, freeCallbacksProvider(&pClientCallbacks));
 
         for(i = 0; i < logIterationCount; ++i) {
-            SPRINTF(filePath, TEST_TEMP_DIR_PATH "kvsProducerLog.%u", i);
+            SPRINTF(filePath, TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".%u", i);
             EXPECT_EQ(STATUS_SUCCESS, fileExists(filePath, &fileFound));
             // only log file with index from 6 to 11 should remain. The rest are rotated out.
             if (i < logIterationCount && i >= (logIterationCount - maxLogFileCount)) {
@@ -166,10 +166,10 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
                                                                          &pClientCallbacks));
 
         // make sure the files dont exist
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLogIndex");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
 
         ULLTOSTR(128, fileIndexBuffer, ARRAY_SIZE(fileIndexBuffer), 10, &fileIndexStrSize);
-        EXPECT_EQ(STATUS_SUCCESS, writeFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), TRUE, FALSE, (PBYTE) fileIndexBuffer, STRLEN(fileIndexBuffer) * SIZEOF(CHAR)));
+        EXPECT_EQ(STATUS_SUCCESS, writeFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, FALSE, (PBYTE) fileIndexBuffer, STRLEN(fileIndexBuffer) * SIZEOF(CHAR)));
 
         EXPECT_EQ(STATUS_SUCCESS, addFileLoggerPlatformCallbacksProvider(pClientCallbacks, MIN_FILE_LOGGER_STRING_BUFFER_SIZE, 5, TEST_TEMP_DIR_PATH_NO_ENDING_SEPARTOR, FALSE));
         logFunc = pClientCallbacks->logPrintFn;
@@ -179,19 +179,19 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
 
         // check that log file name follows last index
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.128"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".128"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
         EXPECT_EQ(TRUE, fileFound); // index file should be there
 
         EXPECT_EQ(STATUS_SUCCESS, freeCallbacksProvider(&pClientCallbacks));
 
         // remaining log get flushed
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.129"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".129"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), TRUE, NULL, &fileIndexBufferSize));
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLogIndex"), TRUE, (PBYTE) fileIndexBuffer, &fileIndexBufferSize));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, NULL, &fileIndexBufferSize));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, (PBYTE) fileIndexBuffer, &fileIndexBufferSize));
         fileIndexBuffer[fileIndexBufferSize] = '\0';
         STRTOUI64(fileIndexBuffer, NULL, 10, &currentFileIndex);
         EXPECT_EQ(130, currentFileIndex); // index should be incremented 3 times
@@ -220,8 +220,8 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
                                                                          &pClientCallbacks));
 
         // make sure the files dont exist
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLogIndex");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.0");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
 
         MEMSET(logMessage, 'a', logMessageSize);
         logMessage[logMessageSize] = '\0';
@@ -239,11 +239,11 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
         EXPECT_EQ(STATUS_SUCCESS, freeCallbacksProvider(&pClientCallbacks));
 
         // no log should be produced as the only log message was dropped because it was too long.
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), TRUE, NULL, &fileBufferLen));
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), TRUE, (PBYTE) fileBuffer, &fileBufferLen));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), TRUE, NULL, &fileBufferLen));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), TRUE, (PBYTE) fileBuffer, &fileBufferLen));
         fileBuffer[fileBufferLen] = '\0';
 
         EXPECT_EQ('a', fileBuffer[STRLEN(fileBuffer) - 1]); // there should be no b in the log.
@@ -275,11 +275,11 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
                                                                          &pClientCallbacks));
 
         // make sure the files dont exist
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLogIndex");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.0");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.1");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.2");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.3");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".3");
 
         MEMSET(logMessage, 'a', logMessageSize);
         logMessage[logMessageSize] = '\0';
@@ -289,29 +289,29 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
 
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.1"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
         MEMSET(logMessage, 'b', logMessageSize);
         logMessage[logMessageSize] = '\0';
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.2"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
 
         // kvsProducerLog.0 is deleted
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(FALSE, fileFound);
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.3"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".3"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.3"), TRUE, NULL, &fileBufferLen));
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.3"), TRUE, (PBYTE) fileBuffer, &fileBufferLen));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".3"), TRUE, NULL, &fileBufferLen));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".3"), TRUE, (PBYTE) fileBuffer, &fileBufferLen));
         fileBuffer[fileBufferLen] = '\0';
         // skip over the timestamp string
         // use STRNCMP to not catch the newline at the end of fileBuffer
@@ -346,10 +346,10 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
                                                                          &pClientCallbacks));
 
         // make sure the files dont exist
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLogIndex");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.0");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.1");
-        FREMOVE(TEST_TEMP_DIR_PATH "kvsProducerLog.2");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2");
 
         MEMSET(logMessage, 'a', logMessageSize);
         logMessage[logMessageSize] = '\0';
@@ -364,18 +364,18 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
 
         logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
 
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
         EXPECT_EQ(STATUS_SUCCESS, freeCallbacksProvider(&pClientCallbacks));
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.0"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
         EXPECT_EQ(FALSE, fileFound);
 
-        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.1"), &fileFound));
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1"), &fileFound));
         EXPECT_EQ(TRUE, fileFound);
 
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.1"), TRUE, NULL, &fileBufferLen));
-        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH "kvsProducerLog.1"), TRUE, (PBYTE) fileBuffer, &fileBufferLen));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1"), TRUE, NULL, &fileBufferLen));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1"), TRUE, (PBYTE) fileBuffer, &fileBufferLen));
         fileBuffer[fileBufferLen] = '\0';
 
         // skip over the timestamp string
@@ -384,6 +384,236 @@ namespace com { namespace amazonaws { namespace kinesis { namespace video {
 
         MEMFREE(logMessage);
         MEMFREE(fileBuffer);
+    }
+
+    TEST_F(FileLoggerFunctionalityTest, basicCommonFileLoggerUsageViaCallbackNoGlobalLog)
+    {
+        PClientCallbacks pClientCallbacks = NULL;
+        UINT32 logMessageSize = MIN_FILE_LOGGER_STRING_BUFFER_SIZE / 2;
+        PCHAR logMessage = (PCHAR) MEMALLOC(logMessageSize + 1);
+        BOOL fileFound = FALSE;
+        LogPrintFunc logFunc;
+        UINT64 currentFileIndex = 0;
+        CHAR fileIndexBuffer[256];
+        UINT64 fileIndexBufferSize = ARRAY_SIZE(fileIndexBuffer);
+
+        MEMSET(logMessage, 'a', logMessageSize);
+        logMessage[logMessageSize] = '\0';
+
+        // make sure the files dont exist
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2");
+
+        EXPECT_EQ(STATUS_SUCCESS, createFileLogger(MIN_FILE_LOGGER_STRING_BUFFER_SIZE, 5, TEST_TEMP_DIR_PATH_NO_ENDING_SEPARTOR, FALSE, FALSE, &logFunc));
+
+        logFunc(LOG_LEVEL_VERBOSE, NULL, (PCHAR) "%s", logMessage);
+        logFunc(LOG_LEVEL_VERBOSE, NULL, (PCHAR) "%s", logMessage);
+        // low log level logs have no effect
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(FALSE, fileFound);
+
+        logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
+        // log should not have been flushed because we havent fill out string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(FALSE, fileFound);
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(FALSE, fileFound); // index file should not be there
+
+        logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
+        // log should have been flushed because the new message overflows string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound);
+
+        // second log is still in buffer. Thus this log will cause the second log to be flushed.
+        logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
+        // log should have been flushed
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME".1"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound);
+
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(TRUE, fileFound); // index file should be there
+
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, NULL, &fileIndexBufferSize));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, (PBYTE) fileIndexBuffer, &fileIndexBufferSize));
+        fileIndexBuffer[fileIndexBufferSize] = '\0';
+        STRTOUI64(fileIndexBuffer, NULL, 10, &currentFileIndex);
+        EXPECT_EQ(2, currentFileIndex); // index should be 2 since we flushed twice.
+
+        EXPECT_EQ(STATUS_SUCCESS, freeFileLogger());
+
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound); // remaining log get flushed when callbacks are freed.
+
+        MEMFREE(logMessage);
+    }
+
+    TEST_F(FileLoggerFunctionalityTest, basicCommonFileLoggerUsageViaCallbackWithGlobalLog)
+    {
+        PClientCallbacks pClientCallbacks = NULL;
+        UINT32 logMessageSize = MIN_FILE_LOGGER_STRING_BUFFER_SIZE / 2;
+        PCHAR logMessage = (PCHAR) MEMALLOC(logMessageSize + 1);
+        BOOL fileFound = FALSE;
+        LogPrintFunc logFunc;
+        UINT64 currentFileIndex = 0;
+        CHAR fileIndexBuffer[256];
+        UINT64 fileIndexBufferSize = ARRAY_SIZE(fileIndexBuffer);
+
+        MEMSET(logMessage, 'a', logMessageSize);
+        logMessage[logMessageSize] = '\0';
+
+        // make sure the files dont exist
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2");
+
+        EXPECT_EQ(STATUS_SUCCESS, createFileLogger(MIN_FILE_LOGGER_STRING_BUFFER_SIZE, 5, TEST_TEMP_DIR_PATH_NO_ENDING_SEPARTOR, FALSE, TRUE, &logFunc));
+
+        logFunc(LOG_LEVEL_VERBOSE, NULL, (PCHAR) "%s", logMessage);
+        logFunc(LOG_LEVEL_VERBOSE, NULL, (PCHAR) "%s", logMessage);
+        // low log level logs have no effect
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(FALSE, fileFound);
+
+        logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
+        // log should not have been flushed because we havent fill out string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(FALSE, fileFound);
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(FALSE, fileFound); // index file should not be there
+
+        logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
+        // log should have been flushed because the new message overflows string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound);
+
+        // second log is still in buffer. Thus this log will cause the second log to be flushed.
+        logFunc(LOG_LEVEL_ERROR, NULL, (PCHAR) "%s", logMessage);
+        // log should have been flushed
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME".1"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound);
+
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(TRUE, fileFound); // index file should be there
+
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, NULL, &fileIndexBufferSize));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, (PBYTE) fileIndexBuffer, &fileIndexBufferSize));
+        fileIndexBuffer[fileIndexBufferSize] = '\0';
+        STRTOUI64(fileIndexBuffer, NULL, 10, &currentFileIndex);
+        EXPECT_EQ(2, currentFileIndex); // index should be 2 since we flushed twice.
+
+        EXPECT_EQ(STATUS_SUCCESS, freeFileLogger());
+
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound); // remaining log get flushed when callbacks are freed.
+
+        MEMFREE(logMessage);
+    }
+
+    TEST_F(FileLoggerFunctionalityTest, basicCommonFileLoggerUsageViaGlobalLog)
+    {
+        PClientCallbacks pClientCallbacks = NULL;
+        UINT32 logMessageSize = MIN_FILE_LOGGER_STRING_BUFFER_SIZE / 2;
+        PCHAR logMessage = (PCHAR) MEMALLOC(logMessageSize + 1);
+        BOOL fileFound = FALSE;
+        UINT64 currentFileIndex = 0;
+        CHAR fileIndexBuffer[256];
+        UINT64 fileIndexBufferSize = ARRAY_SIZE(fileIndexBuffer);
+
+        MEMSET(logMessage, 'a', logMessageSize);
+        logMessage[logMessageSize] = '\0';
+
+        // make sure the files dont exist
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".1");
+        FREMOVE(TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2");
+
+        EXPECT_EQ(STATUS_SUCCESS, createFileLogger(MIN_FILE_LOGGER_STRING_BUFFER_SIZE, 5, TEST_TEMP_DIR_PATH_NO_ENDING_SEPARTOR, FALSE, TRUE, NULL));
+
+        DLOGV("%s", logMessage);
+        DLOGV("%s", logMessage);
+        // low log level logs have no effect
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(FALSE, fileFound);
+
+        DLOGE("%s", logMessage);
+        // log should not have been flushed because we havent fill out string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(FALSE, fileFound);
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(FALSE, fileFound); // index file should not be there
+
+        DLOGE("%s", logMessage);
+        // log should have been flushed because the new message overflows string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound);
+
+        // second log is still in buffer. Thus this log will cause the second log to be flushed.
+        DLOGE("%s", logMessage);
+        // log should have been flushed
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME".1"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound);
+
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(TRUE, fileFound); // index file should be there
+
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, NULL, &fileIndexBufferSize));
+        EXPECT_EQ(STATUS_SUCCESS, readFile((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), TRUE, (PBYTE) fileIndexBuffer, &fileIndexBufferSize));
+        fileIndexBuffer[fileIndexBufferSize] = '\0';
+        STRTOUI64(fileIndexBuffer, NULL, 10, &currentFileIndex);
+        EXPECT_EQ(2, currentFileIndex); // index should be 2 since we flushed twice.
+
+        EXPECT_EQ(STATUS_SUCCESS, freeFileLogger());
+
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (TEST_TEMP_DIR_PATH FILE_LOGGER_LOG_FILE_NAME ".2"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound); // remaining log get flushed when callbacks are freed.
+
+        MEMFREE(logMessage);
+    }
+
+    TEST_F(FileLoggerFunctionalityTest, basicCommonFileLoggerUsageViaGlobalLogSetResetMacros)
+    {
+        PClientCallbacks pClientCallbacks = NULL;
+        UINT32 logMessageSize = MIN_FILE_LOGGER_STRING_BUFFER_SIZE / 2;
+        PCHAR logMessage = (PCHAR) MEMALLOC(logMessageSize + 1);
+        BOOL fileFound = FALSE;
+        UINT64 currentFileIndex = 0;
+        CHAR fileIndexBuffer[256];
+        UINT64 fileIndexBufferSize = ARRAY_SIZE(fileIndexBuffer);
+
+        MEMSET(logMessage, 'a', logMessageSize);
+        logMessage[logMessageSize] = '\0';
+
+        // make sure the files dont exist
+        FREMOVE(FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME);
+        FREMOVE(FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LOG_FILE_NAME ".0");
+        FREMOVE(FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LOG_FILE_NAME ".1");
+        FREMOVE(FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LOG_FILE_NAME ".2");
+
+        EXPECT_EQ(STATUS_SUCCESS, SET_FILE_LOGGER());
+
+        DLOGE("%s", logMessage);
+        // log should not have been flushed because we havent fill out string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(FALSE, fileFound);
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(FALSE, fileFound); // index file should not be there
+
+        EXPECT_EQ(STATUS_SUCCESS, RESET_FILE_LOGGER());
+
+        // log should not have been flushed because we havent fill out string buffer
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LOG_FILE_NAME ".0"), &fileFound));
+        EXPECT_EQ(TRUE, fileFound); // remaining log get flushed when callbacks are freed.
+        EXPECT_EQ(STATUS_SUCCESS, fileExists((PCHAR) (FILE_LOGGER_LOG_FILE_DIRECTORY_PATH FILE_LOGGER_LAST_INDEX_FILE_NAME), &fileFound));
+        EXPECT_EQ(TRUE, fileFound);
+
+        MEMFREE(logMessage);
+
+        // Ensure we can still log err
+        DLOGE("Printout after reset");
     }
 }
 }
